Rename data -> object
