+++ moq-transport/src/lib.rs
@@ -1,6 +1,6 @@
 pub mod coding;
 pub mod control;
-pub mod data;
+pub mod object;
 pub mod server;
 pub mod setup;
